Add files via upload
In this project, I uploaded an old table that I had created named Products. I referenced the table to get a replicate and renamed it Products and colors just to create a mapping with the columns to create a relationship between them. I clicked on choose columns to select the columns needed and after that loaded to a connection and into the data model and created the relationship in diagram view.